fix(bundles): apply cross-provider review — and two defects it surfaced
Review (sol@high) returned two HIGH, one MEDIUM, one LOW. All four
addressed; the two HIGHs were about what the caller gets charged for.

prospect-brief over-claimed. "Tell me about their company" applied to
person@gmail.com returned Google's mail setup, domain reputation and tech
stack as if it were the prospect's employer — misleading output the
caller paid EUR0.30 for, which is worse than a failed call. There is no
conditional-step support to short-circuit on, so the fix is to stop
claiming it: the bundle now describes profiling the domain behind an
address, and says outright that a consumer mailbox returns that
provider's setup, pointing at the is_disposable and is_role_address flags
step 1 already returns.

page-seo-check ran all four steps concurrently, so an unreachable page
burned four calls. The platform refunds only when EVERY step fails, so a
partial failure — url-health-check succeeding with is_up=false while the
other three fail — bills in full. url-health-check now runs alone and
first, which guarantees the availability answer is in the result rather
than racing three slower checks that will all fail. Skip-and-refund on an
unreachable target needs conditional steps and is a platform gap, not
something a bundle definition can fix; recorded in the PR.

The catalogue is now a pure data module. The entry-point guard stopped
seeding on import, but the reviewer noted importing still ran dotenv and
populated DATABASE_URL, so a test process that had deliberately unset it
got a live connection string anyway — and the test named "side-effect
free" passed regardless. solution-catalogue.ts imports nothing that
reaches the environment or the database.

The execution-order test compared $steps[N] against array position.
$steps[N] resolves by sorted stepOrder, so reordering the array while
leaving stepOrder alone could slip a forward reference past it. Now
sorted by stepOrder, with contiguous-numbering and same-parallel-group
assertions alongside.

That strengthened test immediately caught a second pre-existing defect:
dependency-risk-check had license-compatibility-check reading
$steps[0].license.spdx while sharing parallel group 1 with the step it
reads from. Same group runs concurrently, so the executor's null fallback
could silently drop the licence list and check nothing. Moved to its own
group. Unlike the token-project-dd duplicate, this one is unambiguous —
a dependent step must not race its source, no pricing judgement involved
— and the bundle has no external sales, so nothing regressed.

Full suite 1664 green; 17 bundle tests; typecheck, PII, cost-class,
SSRF-inventory and bare-catch gates clean.
