+++ apps/api/src/db/solution-catalogue.ts
@@ -1,0 +1,3062 @@
+/**
+ * The bundle catalogue — data only.
+ *
+ * Split out of seed-solutions.ts on 2026-08-16 so the definitions can be
+ * imported, validated and tested without touching the environment or the
+ * database. The seed script previously declared the catalogue AND seeded it on
+ * import: `seed()` ran unconditionally at module load, so any import wrote to
+ * whatever DATABASE_URL pointed at. A throwaway validation script imported it
+ * against production and only failed to seed because the specifier happened
+ * not to resolve.
+ *
+ * An entry-point guard in the seed script fixed the write. This file fixes the
+ * rest: cross-provider review pointed out that importing the catalogue still
+ * ran dotenv and populated DATABASE_URL as a side effect, so a test process
+ * that had deliberately unset it got a live connection string anyway. Nothing
+ * in this module imports dotenv, the database, or anything that reaches them.
+ */
+
+// ─── Solution definitions ───────────────────────────────────────────────────
+
+export interface SolutionStep {
+  capabilitySlug: string;
+  stepOrder: number;
+  canParallel: boolean;
+  parallelGroup: number | null;
+  inputMap: Record<string, string>;
+}
+
+export interface SolutionDef {
+  slug: string;
+  name: string;
+  marketingName: string;
+  description: string;
+  longDescription?: string;
+  agentDescription?: string;
+  category: string;
+  priceCents: number;
+  componentSumCents: number;
+  valueTier: string;
+  maintenanceLevel: string;
+  geography: string;
+  targetAudience: string;
+  transparencyTag: string | null;
+  extendsWith: string[];
+  inputSchema: Record<string, unknown>;
+  exampleInput?: Record<string, unknown>;
+  exampleOutput?: Record<string, unknown>;
+  steps: SolutionStep[];
+}
+
+// ─── Compliance coverage builder ────────────────────────────────────────────
+
+
+/**
+ * Exported so the definitions can be validated without running the seed.
+ *
+ * Until 2026-08-16 this module both declared the catalogue AND seeded it on
+ * import — `seed()` was called at the bottom unconditionally, so any `import`
+ * of this file wrote to whatever database DATABASE_URL pointed at. A
+ * throwaway validation script imported it while writing the growth bundles
+ * and only failed to seed production because the specifier happened not to
+ * resolve. The export plus the entry-point guard below turn that from luck
+ * into a property of the file.
+ */
+export const SOLUTIONS: SolutionDef[] = [
+  // ── 1. Nordic KYC — Sweden ──
+  {
+    slug: "kyc-sweden",
+    name: "Nordic KYC — Sweden",
+    marketingName: "Nordic KYC — Sweden",
+    description:
+      "Check if a Swedish company is real and safe to do business with. Official registry data, VAT validation, sanctions screening, PEP check, and adverse media scan in one call.",
+    longDescription:
+      "Looks up the company in Sweden's official business registry (Bolagsverket), validates their EU VAT number against VIES, and screens the company name against OFAC, EU, and UN sanctions lists. Use this before onboarding a customer, signing a contract, or generating a partnership agreement involving a Swedish entity.",
+    agentDescription:
+      "verify swedish company, is this swedish company legit, KYC check sweden, check swedish org number, onboard swedish business customer, validate swedish organization, bolagsverket lookup",
+    category: "compliance-verification",
+    priceCents: 150,
+    componentSumCents: 145,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "nordic",
+    targetAudience: "Developers building fintech tools, onboarding flows, or any agent that verifies Nordic businesses",
+    transparencyTag: "mixed",
+    extendsWith: ["company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        org_number: {
+          type: "string",
+          description: "Swedish organization number",
+        },
+      },
+      required: ["org_number"],
+    },
+    exampleInput: { org_number: "556703-7485" },
+    exampleOutput: {
+      company_name: "Spotify AB",
+      org_number: "556703-7485",
+      vat_valid: true,
+      vat_number: "SE556703748501",
+      is_sanctioned: false,
+      match_count: 0,
+      is_pep: false,
+      pep_match_count: 0,
+      adverse_media_risk: "none",
+    },
+    steps: [
+      {
+        capabilitySlug: "swedish-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { org_number: "$input.org_number" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$steps[0].vat_number" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+    ],
+  },
+
+  // ── 2. Nordic KYC — Norway ──
+  {
+    slug: "kyc-norway",
+    name: "Nordic KYC — Norway",
+    marketingName: "Nordic KYC — Norway",
+    description:
+      "Check if a Norwegian company is real and safe to do business with. Official registry data, VAT validation, sanctions screening, PEP check, and adverse media scan in one call.",
+    longDescription:
+      "Looks up the company in Norway's Brønnøysund Register Centre, validates their VAT number, and screens against international sanctions lists. Use this before doing business with a Norwegian entity.",
+    agentDescription:
+      "verify norwegian company, is this norwegian company legit, KYC check norway, check norwegian org number, brønnøysund lookup, validate norwegian organization",
+    category: "compliance-verification",
+    priceCents: 150,
+    componentSumCents: 145,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "nordic",
+    targetAudience: "Developers building fintech tools or agents that verify Norwegian businesses",
+    transparencyTag: "mixed",
+    extendsWith: ["domain-reputation", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        org_number: {
+          type: "string",
+          description: "Norwegian organization number",
+        },
+      },
+      required: ["org_number"],
+    },
+    exampleInput: { org_number: "923609016" },
+    exampleOutput: {
+      company_name: "EQUINOR ASA",
+      org_number: "923609016",
+      vat_valid: true,
+      vat_number: "NO923609016MVA",
+      is_sanctioned: false,
+      match_count: 0,
+      is_pep: false,
+      pep_match_count: 0,
+      adverse_media_risk: "none",
+    },
+    steps: [
+      {
+        capabilitySlug: "norwegian-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { org_number: "$input.org_number" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$steps[0].vat_number" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+    ],
+  },
+
+  // ── 3. Nordic KYC — Denmark ──
+  {
+    slug: "kyc-denmark",
+    name: "Nordic KYC — Denmark",
+    marketingName: "Nordic KYC — Denmark",
+    description:
+      "Check if a Danish company is real and safe to do business with. Official registry data, VAT validation, sanctions screening, PEP check, and adverse media scan in one call.",
+    longDescription:
+      "Looks up the company in Denmark's CVR register, validates their EU VAT number, and screens against international sanctions lists. Use this before doing business with a Danish entity.",
+    agentDescription:
+      "verify danish company, is this danish company legit, KYC check denmark, check danish CVR number, validate danish organization",
+    category: "compliance-verification",
+    priceCents: 150,
+    componentSumCents: 145,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "nordic",
+    targetAudience: "Developers building fintech tools or agents that verify Danish businesses",
+    transparencyTag: "mixed",
+    extendsWith: ["domain-reputation", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        cvr_number: { type: "string", description: "Danish CVR number" },
+      },
+      required: ["cvr_number"],
+    },
+    exampleInput: { cvr_number: "47458714" },
+    exampleOutput: {
+      company_name: "LEGO System A/S",
+      cvr_number: "47458714",
+      vat_valid: true,
+      vat_number: "DK47458714",
+      is_sanctioned: false,
+      match_count: 0,
+      is_pep: false,
+      pep_match_count: 0,
+      adverse_media_risk: "none",
+    },
+    steps: [
+      {
+        capabilitySlug: "danish-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { cvr_number: "$input.cvr_number" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$steps[0].vat_number" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+    ],
+  },
+
+  // ── 4. Nordic KYC — Finland ──
+  {
+    slug: "kyc-finland",
+    name: "Nordic KYC — Finland",
+    marketingName: "Nordic KYC — Finland",
+    description:
+      "Check if a Finnish company is real and safe to do business with. Official registry data, VAT validation, sanctions screening, PEP check, and adverse media scan in one call.",
+    longDescription:
+      "Looks up the company in Finland's PRH business register using their Y-tunnus (business ID), validates their EU VAT number, and screens against international sanctions lists. Use this before doing business with a Finnish entity.",
+    agentDescription:
+      "verify finnish company, is this finnish company legit, KYC check finland, check finnish business id, Y-tunnus lookup, validate finnish organization",
+    category: "compliance-verification",
+    priceCents: 150,
+    componentSumCents: 145,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "nordic",
+    targetAudience: "Developers building fintech tools or agents that verify Finnish businesses",
+    transparencyTag: "mixed",
+    extendsWith: ["domain-reputation", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        business_id: {
+          type: "string",
+          description: "Finnish business ID (Y-tunnus)",
+        },
+      },
+      required: ["business_id"],
+    },
+    exampleInput: { business_id: "0112038-9" },
+    exampleOutput: {
+      company_name: "Nokia Oyj",
+      business_id: "0112038-9",
+      vat_valid: true,
+      vat_number: "FI01120389",
+      is_sanctioned: false,
+      match_count: 0,
+      is_pep: false,
+      pep_match_count: 0,
+      adverse_media_risk: "none",
+    },
+    steps: [
+      {
+        capabilitySlug: "finnish-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { business_id: "$input.business_id" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$steps[0].vat_number" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+    ],
+  },
+
+  // ── 5. B2B Payment Validation ──
+  {
+    slug: "payment-validate",
+    name: "B2B Payment Validation",
+    marketingName: "B2B Payment Validation",
+    description:
+      "Verify bank details and tax IDs before your agent processes a payment. IBAN validation, bank identification, and counterparty VAT check.",
+    longDescription:
+      "Validates the IBAN structure and identifies the receiving bank, then confirms the counterparty's VAT registration is active. Catches invalid bank details and unregistered VAT numbers before money moves. Use this in any payment flow, invoice processing, or vendor onboarding.",
+    agentDescription:
+      "validate IBAN before payment, check bank details, verify payment recipient, is this IBAN real, validate counterparty before transfer, B2B payment check",
+    category: "finance-banking",
+    priceCents: 25,
+    componentSumCents: 15,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "eu",
+    targetAudience: "Developers building payment flows or B2B finance tools",
+    transparencyTag: "algorithmic",
+    extendsWith: ["exchange-rate", "sanctions-check", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        iban: { type: "string" },
+        vat_number: { type: "string" },
+      },
+      required: ["iban", "vat_number"],
+    },
+    exampleInput: {
+      iban: "DE89370400440532013000",
+      vat_number: "DE136695976",
+    },
+    exampleOutput: {
+      iban_valid: true,
+      country_code: "DE",
+      bank_code: "37040044",
+      vat_valid: true,
+      vat_company_name: "SAP SE",
+    },
+    steps: [
+      {
+        capabilitySlug: "iban-validate",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { iban: "$input.iban" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { vat_number: "$input.vat_number" },
+      },
+    ],
+  },
+
+  // ── 6. SEPA Payment Readiness ──
+  {
+    slug: "sepa-readiness",
+    name: "SEPA Payment Readiness",
+    marketingName: "SEPA Payment Readiness Check",
+    description:
+      "Everything your agent needs before initiating a SEPA transfer: IBAN validation, bank identification, VAT verification, and current exchange rate.",
+    longDescription:
+      "Validates the receiving IBAN, identifies the bank, confirms the counterparty's VAT registration, and fetches the current exchange rate if the source currency isn't EUR. One call replaces four separate API integrations. Use before any European bank transfer.",
+    agentDescription:
+      "prepare SEPA payment, SEPA transfer readiness check, European bank transfer validation, verify before SEPA transfer, IBAN and VAT for SEPA",
+    category: "finance-banking",
+    priceCents: 30,
+    componentSumCents: 17,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "eu",
+    targetAudience: "Developers building European payment flows or SEPA integrations",
+    transparencyTag: "algorithmic",
+    extendsWith: ["sanctions-check", "bank-bic-lookup", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        iban: { type: "string" },
+        vat_number: { type: "string" },
+        source_currency: {
+          type: "string",
+          description: "ISO 4217 currency code, omit if EUR",
+        },
+      },
+      required: ["iban", "vat_number"],
+    },
+    exampleInput: {
+      iban: "DE89370400440532013000",
+      vat_number: "DE136695976",
+      source_currency: "SEK",
+    },
+    exampleOutput: {
+      iban_valid: true,
+      country_code: "DE",
+      bank_code: "37040044",
+      vat_valid: true,
+      vat_company_name: "SAP SE",
+      exchange_rate: 0.0936,
+      exchange_rate_date: "2026-03-05",
+    },
+    steps: [
+      {
+        capabilitySlug: "iban-validate",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { iban: "$input.iban" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { vat_number: "$input.vat_number" },
+      },
+      {
+        capabilitySlug: "exchange-rate",
+        stepOrder: 3,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { from: "$input.source_currency", to: "EUR" },
+      },
+    ],
+  },
+
+  // ── 7. Lead Email Verify ──
+  {
+    slug: "lead-email-verify",
+    name: "Lead Email Verify",
+    marketingName: "Lead Email Verification",
+    description:
+      "Your agent found a prospect's email — is it real? Deliverability check, DNS, and domain reputation in one call.",
+    longDescription:
+      "Validates the email address for deliverability, checks the domain's DNS configuration and MX records, and scores the domain's reputation. Catches disposable addresses, invalid mailboxes, and suspicious domains before your agent sends outreach. Use in any sales, marketing, or CRM enrichment pipeline.",
+    agentDescription:
+      "is this email real, check email before sending, validate prospect email, email deliverability check, verify email address, is this a real mailbox",
+    category: "sales-outreach",
+    priceCents: 20,
+    componentSumCents: 11,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "us-global",
+    targetAudience:
+      "Developers building outbound sales agents, email verification, or lead gen pipelines",
+    transparencyTag: "algorithmic",
+    extendsWith: ["company-enrich", "tech-stack-detect", "whois-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: { email: { type: "string", format: "email" } },
+      required: ["email"],
+    },
+    exampleInput: { email: "test@google.com" },
+    exampleOutput: {
+      valid: true,
+      domain: "google.com",
+      has_mx: true,
+      is_disposable: false,
+      reputation_score: 92,
+    },
+    steps: [
+      {
+        capabilitySlug: "email-validate",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { email: "$input.email" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+    ],
+  },
+
+  // ── 8. US Company Verify ──
+  {
+    slug: "verify-us-company",
+    name: "US Company Verify",
+    marketingName: "US Company Verification",
+    description:
+      "Look up any US company using SEC EDGAR data, screen against international sanctions lists, check for politically exposed persons (PEP), and scan for adverse media coverage. Filing data, company status, sanctions, PEP, and media check.",
+    longDescription:
+      "Queries SEC EDGAR for official filing data using a company name, ticker symbol, or CIK number, then screens the company against OFAC, EU, and UN sanctions lists. Returns company name, CIK, state of incorporation, and sanctions status. Use before doing business with a US entity.",
+    agentDescription:
+      "verify US company, check if american company exists, SEC company lookup, is this US company legit, US company sanctions check, EDGAR lookup",
+    category: "compliance-verification",
+    priceCents: 130,
+    componentSumCents: 135,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "us",
+    targetAudience:
+      "Developers doing vendor/partner onboarding or compliance checks on US companies",
+    transparencyTag: "ai_generated",
+    extendsWith: ["domain-reputation", "tech-stack-detect", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company: {
+          type: "string",
+          description: "CIK number, ticker symbol, or company name",
+        },
+      },
+      required: ["company"],
+    },
+    exampleInput: { company: "AAPL" },
+    exampleOutput: {
+      company_name: "Apple Inc.",
+      cik: "0000320193",
+      state: "CA",
+      is_sanctioned: false,
+      match_count: 0,
+      is_pep: false,
+      pep_match_count: 0,
+      adverse_media_risk: "none",
+    },
+    steps: [
+      {
+        capabilitySlug: "us-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { company: "$input.company" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+    ],
+  },
+
+  // ── 9. Domain Intelligence ──
+  {
+    slug: "domain-intel",
+    name: "Domain Intelligence",
+    marketingName: "Domain Intelligence Report",
+    description:
+      "Everything your agent needs to evaluate a domain: registration details, DNS setup, SSL health, and reputation score.",
+    longDescription:
+      "Combines WHOIS registration data, DNS record analysis, SSL certificate validation, and domain reputation scoring into one report. Use when your agent is researching a company, vetting a vendor, qualifying a lead, or checking a link before following it.",
+    agentDescription:
+      "look up this domain, who owns this website, is this domain trustworthy, domain background check, domain registration info, check website reputation",
+    category: "security-risk",
+    priceCents: 35,
+    componentSumCents: 21,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "Any developer doing vendor assessment, lead qualification, or security monitoring",
+    transparencyTag: "algorithmic",
+    extendsWith: ["tech-stack-detect", "page-speed-test", "backlink-check"],
+    inputSchema: {
+      type: "object",
+      properties: { domain: { type: "string" } },
+      required: ["domain"],
+    },
+    exampleInput: { domain: "google.com" },
+    exampleOutput: {
+      registrar: "MarkMonitor Inc.",
+      ssl_valid: true,
+      dns_records: true,
+      reputation_score: 95,
+    },
+    steps: [
+      {
+        capabilitySlug: "whois-lookup",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── 10. Web Extract & Clean ──
+  {
+    slug: "web-extract-clean",
+    name: "Web Extract & Clean",
+    marketingName: "Web Extract & Clean",
+    description:
+      "Fetch any web page — even behind anti-bot walls — convert to clean markdown, and strip PII automatically. Ready for your RAG pipeline.",
+    longDescription:
+      "Uses a headless browser to fetch JavaScript-rendered pages that block direct requests, converts the content to clean markdown, then automatically detects and redacts PII (names, emails, phone numbers, addresses) before returning. Ready to drop into your RAG pipeline, knowledge base, or agent context window.",
+    agentDescription:
+      "scrape this page, get content from URL, extract web page for RAG, page blocked need content, convert URL to markdown, fetch website content, web scraping with anti-bot",
+    category: "data-research",
+    priceCents: 30,
+    componentSumCents: 20,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience:
+      "Any developer building research agents, RAG pipelines, or content ingestion",
+    transparencyTag: "mixed",
+    extendsWith: ["structured-scrape", "translate", "summarize"],
+    inputSchema: {
+      type: "object",
+      properties: { url: { type: "string", format: "uri" } },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://example.com" },
+    exampleOutput: {
+      markdown: "# Example Domain\n\nThis domain is for use in illustrative examples...",
+      pii_redacted: 0,
+      word_count: 58,
+    },
+    steps: [
+      {
+        capabilitySlug: "url-to-markdown",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "pii-redact",
+        stepOrder: 2,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { text: "$steps[0].markdown" },
+      },
+    ],
+  },
+
+  // ── 11. Email Deliverability Audit ──
+  {
+    slug: "email-audit",
+    name: "Email Deliverability Audit",
+    marketingName: "Email Deliverability Audit",
+    description:
+      "Will emails to this domain actually arrive? SPF, DKIM, DMARC, MX records, SSL, and blacklist status — the complete deliverability audit.",
+    longDescription:
+      "Checks every factor that determines inbox placement: SPF record configuration, DKIM signing, DMARC policy, MX record health, SSL certificate status, and blacklist presence. Use when evaluating email infrastructure for lead qualification, vendor assessment, or monitoring your own domain's deliverability.",
+    agentDescription:
+      "check email config for domain, will emails arrive at this domain, email infrastructure audit, SPF DKIM DMARC check, email deliverability test, can this domain receive email",
+    category: "security-risk",
+    priceCents: 25,
+    componentSumCents: 16,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience: "Any developer or team sending email at scale",
+    transparencyTag: "algorithmic",
+    extendsWith: ["domain-reputation", "whois-lookup", "tech-stack-detect"],
+    inputSchema: {
+      type: "object",
+      properties: { domain: { type: "string" } },
+      required: ["domain"],
+    },
+    exampleInput: { domain: "google.com" },
+    exampleOutput: {
+      score: 95,
+      grade: "A",
+      spf: "pass",
+      dmarc: "pass",
+      ssl_valid: true,
+    },
+    steps: [
+      {
+        capabilitySlug: "email-deliverability-check",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── 12. Website Health Check ──
+  {
+    slug: "website-health",
+    name: "Website Health Check",
+    marketingName: "Website Health Check",
+    description:
+      "Quick technical health check on any website: SSL certificate chain, email deliverability configuration, and page load performance.",
+    longDescription:
+      "Validates the full SSL certificate chain, checks email deliverability configuration (SPF/DKIM/DMARC), and measures page load performance. Your DevOps agent's first check when monitoring a service, or a quick sanity check before your agent interacts with an external site.",
+    agentDescription:
+      "is this website healthy, check website performance, site health check, SSL and speed test, website technical audit, is this site working properly",
+    category: "security-risk",
+    priceCents: 40,
+    componentSumCents: 25,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "DevOps teams, B2B SaaS, or agents monitoring external services",
+    transparencyTag: "algorithmic",
+    extendsWith: ["dns-lookup", "whois-lookup", "domain-reputation"],
+    inputSchema: {
+      type: "object",
+      properties: { domain: { type: "string" } },
+      required: ["domain"],
+    },
+    exampleInput: { domain: "google.com" },
+    exampleOutput: {
+      certificate_chain: [{ subject: "*.google.com", issuer: "GTS CA 1C3" }],
+      email_score: 95,
+      performance_score: 92,
+    },
+    steps: [
+      {
+        capabilitySlug: "ssl-certificate-chain",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "email-deliverability-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "page-speed-test",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── 13. EU AI Act Risk Assessment ──
+  {
+    slug: "ai-act-assess",
+    name: "EU AI Act Risk Assessment",
+    marketingName: "EU AI Act Risk Assessment",
+    description:
+      "Classify your AI system's risk level under the EU AI Act before the August 2026 deadline. Risk classification, obligations, and supervisory authority.",
+    longDescription:
+      "Analyzes your AI system description against EU AI Act risk categories, identifies matched articles and specific obligations, looks up the relevant national supervisory authority, and checks for prior enforcement actions in your sector. Use before deploying any AI system in the EU.",
+    agentDescription:
+      "EU AI Act classification, is my AI system high risk, AI compliance check, what are my AI Act obligations, EU AI regulation check, AI risk assessment Europe",
+    category: "legal-regulatory",
+    priceCents: 80,
+    componentSumCents: 45,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "eu",
+    targetAudience:
+      "Any developer deploying AI systems in the EU. August 2026 enforcement deadline.",
+    transparencyTag: "mixed",
+    extendsWith: ["gdpr-website-check", "privacy-policy-analyze", "cookie-scan"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        description: {
+          type: "string",
+          description: "Describe what your AI system does",
+        },
+        deployment_country: {
+          type: "string",
+          description: "ISO 3166-1 alpha-2 country code",
+        },
+        company: { type: "string" },
+      },
+      required: ["description", "deployment_country"],
+    },
+    exampleInput: {
+      description:
+        "AI system that triages patient symptoms and recommends urgency level for GP appointments",
+      deployment_country: "SE",
+      company: "Kry",
+    },
+    exampleOutput: {
+      risk_level: "high",
+      category: "Access to essential services",
+      obligations: [
+        "Conformity assessment",
+        "Risk management system",
+        "Data governance",
+      ],
+      supervisory_authority: { name: "IMY", country: "SE" },
+      prior_enforcement: [],
+    },
+    steps: [
+      {
+        capabilitySlug: "eu-ai-act-classify",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: {
+          description: "$input.description",
+        },
+      },
+      {
+        capabilitySlug: "data-protection-authority-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { country_code: "$input.deployment_country" },
+      },
+      {
+        capabilitySlug: "gdpr-fine-lookup",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company: "$input.company" },
+      },
+    ],
+  },
+
+  // ── 14. Website GDPR Audit ──
+  {
+    slug: "gdpr-audit",
+    name: "Website GDPR Audit",
+    marketingName: "Website GDPR Compliance Audit",
+    description:
+      "Is this website GDPR compliant? Scans cookies, consent mechanisms, privacy policy, SSL, and identifies the relevant data protection authority.",
+    longDescription:
+      "Comprehensive GDPR assessment: scans for cookie consent implementation, analyzes tracking scripts, evaluates privacy policy quality, checks SSL security, and identifies the responsible data protection authority for the website's jurisdiction. Use before your agent ships a feature, evaluates a vendor, or monitors compliance across a portfolio.",
+    agentDescription:
+      "GDPR check this website, is this site GDPR compliant, cookie and privacy audit, check website privacy compliance, European data protection check, does this website follow GDPR",
+    category: "legal-regulatory",
+    priceCents: 100,
+    componentSumCents: 63,
+    valueTier: "compliance",
+    maintenanceLevel: "low-medium",
+    geography: "eu-global",
+    targetAudience:
+      "Any developer assessing vendor/partner compliance or monitoring GDPR across a portfolio",
+    transparencyTag: "mixed",
+    extendsWith: ["tech-stack-detect", "domain-reputation", "dns-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        url: { type: "string", format: "uri" },
+        country_code: {
+          type: "string",
+          description: "ISO 3166-1 alpha-2 (for DPA lookup)",
+        },
+      },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://google.com", country_code: "SE" },
+    exampleOutput: {
+      gdpr_score: 72,
+      grade: "B",
+      has_cookie_consent: true,
+      tracking_scripts: { google_analytics: true },
+      privacy_policy: { data_retention_mentioned: true, dpo_listed: false },
+      ssl_valid: true,
+      supervisory_authority: { name: "IMY" },
+    },
+    steps: [
+      {
+        capabilitySlug: "gdpr-website-check",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "cookie-scan",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "privacy-policy-analyze",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.url" },
+      },
+      {
+        capabilitySlug: "data-protection-authority-lookup",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { country_code: "$input.country_code" },
+      },
+    ],
+  },
+
+  // ── 15. Competitive Intelligence Snapshot ──
+  {
+    slug: "competitor-snapshot",
+    name: "Competitive Intelligence Snapshot",
+    marketingName: "Competitive Intelligence Snapshot",
+    description:
+      "Tech stack, SEO, landing page analysis, social presence — the competitive read your agent can run in seconds.",
+    longDescription:
+      "Detects the competitor's technology stack (frameworks, hosting, analytics), audits their SEO health (meta tags, structure, performance), analyzes landing page conversion elements, and checks social media presence. Use for market research, fundraising prep, GTM positioning, or any agent building competitive intelligence.",
+    agentDescription:
+      "research this competitor, competitive analysis, what tech does this company use, analyze competitor website, competitor intelligence, market research on company",
+    category: "data-research",
+    priceCents: 140,
+    componentSumCents: 110,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience:
+      "Any developer building competitive intelligence, market research, or fundraising prep tools",
+    transparencyTag: "mixed",
+    extendsWith: ["page-speed-test", "backlink-check", "domain-reputation"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        competitor_url: { type: "string", format: "uri" },
+        brand_username: {
+          type: "string",
+          description: "Social media username (optional)",
+        },
+      },
+      required: ["competitor_url"],
+    },
+    exampleInput: {
+      competitor_url: "https://stripe.com",
+      brand_username: "stripe",
+    },
+    exampleOutput: {
+      tech_stack: { frontend: "React", hosting: "AWS" },
+      seo: { overall_score: 82, issues: [] },
+      landing_page: { overall_score: 78, strengths: [], weaknesses: [] },
+      social: { github: true, twitter: true, linkedin: true },
+    },
+    steps: [
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.competitor_url" },
+      },
+      {
+        capabilitySlug: "seo-audit",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.competitor_url" },
+      },
+      {
+        capabilitySlug: "landing-page-roast",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.competitor_url" },
+      },
+      {
+        capabilitySlug: "social-profile-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { username: "$input.brand_username" },
+      },
+    ],
+  },
+  // ── 16. Vendor Risk Assessment ──
+  {
+    slug: "vendor-risk-assess",
+    name: "Vendor Risk Assessment",
+    marketingName: "Vendor Risk Assessment",
+    description:
+      "SEC data, sanctions screening, domain security — answers 'should we do business with this company?' in one call.",
+    longDescription:
+      "Checks a company across 7 data sources in one call: SEC EDGAR, OFAC/EU/UN sanctions, WHOIS, DNS, SSL, domain reputation, and HTTP security headers. All steps run in parallel, quality-verified before delivery. Use before onboarding a vendor, signing a partnership, or approving a procurement request.",
+    agentDescription:
+      "assess vendor risk, should we do business with this company, vendor due diligence, vendor security check, is this vendor safe, company risk assessment, supplier evaluation, partnership due diligence",
+    category: "security-risk",
+    priceCents: 180,
+    componentSumCents: 129,
+    valueTier: "verification",
+    maintenanceLevel: "very-low",
+    geography: "us-global",
+    targetAudience:
+      "Developers building procurement agents, vendor management, or partnership due diligence tools",
+    transparencyTag: "mixed",
+    extendsWith: ["seo-audit", "company-enrich", "tech-stack-detect"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company: {
+          type: "string",
+          description: "Company name, CIK, or ticker",
+        },
+        domain: {
+          type: "string",
+          description: "Company domain to assess",
+        },
+      },
+      required: ["company", "domain"],
+    },
+    exampleInput: { company: "Apple Inc", domain: "apple.com" },
+    exampleOutput: {
+      company_name: "Apple Inc.",
+      cik: "0000320193",
+      is_sanctioned: false,
+      domain_reputation_score: 95,
+      ssl_valid: true,
+      header_security_grade: "A",
+    },
+    steps: [
+      {
+        capabilitySlug: "us-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { company: "$input.company" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { name: "$steps[0].company_name" },
+      },
+      {
+        capabilitySlug: "whois-lookup",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 6,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "header-security-check",
+        stepOrder: 7,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { url: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── 17. Lead Enrichment & Qualification ──
+  {
+    slug: "lead-enrich",
+    name: "Lead Enrichment & Qualification",
+    marketingName: "Lead Enrichment & Qualification",
+    description:
+      "Everything your sales agent needs from one email address — validation, domain intel, reputation, registrar, and tech stack detection.",
+    longDescription:
+      "Starting from just an email address: validates deliverability, checks DNS records, scores domain reputation, looks up WHOIS registration data, and detects the company's technology stack. Returns everything an outbound agent needs to qualify a prospect and personalize outreach. Use in CRM enrichment, SDR automation, or account research pipelines.",
+    agentDescription:
+      "enrich this lead, prospect research from email, lead qualification, what company is this email from, CRM enrichment, sales prospecting data, who is this prospect",
+    category: "sales-outreach",
+    priceCents: 65,
+    componentSumCents: 46,
+    valueTier: "verification",
+    maintenanceLevel: "low",
+    geography: "us-global",
+    targetAudience:
+      "Developers building outbound sales agents, CRM enrichment, or account research tools",
+    transparencyTag: "mixed",
+    extendsWith: ["company-enrich", "seo-audit", "landing-page-roast"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        email: {
+          type: "string",
+          format: "email",
+          description: "Prospect email address",
+        },
+      },
+      required: ["email"],
+    },
+    exampleInput: { email: "jane@acme.com" },
+    exampleOutput: {
+      valid: true,
+      domain: "acme.com",
+      has_mx: true,
+      reputation_score: 78,
+      registrar: "GoDaddy",
+      tech_stack: { frontend: "React", cms: "WordPress" },
+    },
+    steps: [
+      {
+        capabilitySlug: "email-validate",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { email: "$input.email" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "whois-lookup",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$steps[0].domain" },
+      },
+    ],
+  },
+
+  // ── 18. Website Security Audit ──
+  {
+    slug: "website-security-audit",
+    name: "Website Security Audit",
+    marketingName: "Website Security Audit",
+    description:
+      "Is this website secure? SSL, HTTP headers, DNS, and tech stack — the security audit your agent runs before trusting any URL.",
+    longDescription:
+      "Comprehensive security posture assessment: validates SSL certificate health, audits HTTP security header configuration (CSP, HSTS, X-Frame-Options), checks DNS security, and maps the technology surface area. Use when your security agent evaluates a vendor, your procurement agent vets a SaaS tool, or your monitoring agent checks your own infrastructure.",
+    agentDescription:
+      "security audit this website, is this URL safe, check website security, SSL and header security check, website vulnerability assessment, is this site secure",
+    category: "security-risk",
+    priceCents: 45,
+    componentSumCents: 29,
+    valueTier: "verification",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience:
+      "Security teams, developers filling security questionnaires, or compliance monitoring agents",
+    transparencyTag: "mixed",
+    extendsWith: ["ssl-certificate-chain", "page-speed-test", "whois-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        url: {
+          type: "string",
+          description: "URL to audit",
+        },
+      },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://example.com" },
+    exampleOutput: {
+      ssl_valid: true,
+      header_security_grade: "B",
+      dns_records: true,
+      tech_stack: { frontend: "Vanilla", hosting: "AWS" },
+    },
+    steps: [
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.url" },
+      },
+      {
+        capabilitySlug: "header-security-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.url" },
+      },
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+    ],
+  },
+
+  // ── 19. Prospect Company Profile ──
+  {
+    slug: "prospect-profile",
+    name: "Prospect Company Profile",
+    marketingName: "Prospect Company Profile",
+    description:
+      "Everything your sales agent researches before a call: SEC data, tech stack, web presence, and domain credibility — delivered in seconds.",
+    longDescription:
+      "Given a company name and URL: pulls SEC EDGAR filing data, detects their technology stack, audits their SEO and web presence strength, and scores their domain credibility. The complete pre-call research package that would take a human 30 minutes, delivered to your agent in seconds.",
+    agentDescription:
+      "research this prospect, company profile for sales call, what does this company do, pre-call research, prospect intelligence, account research before meeting",
+    category: "sales-outreach",
+    priceCents: 180,
+    componentSumCents: 140,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "us-global",
+    targetAudience:
+      "Developers building account research, CRM enrichment, or pre-call intelligence tools",
+    transparencyTag: "mixed",
+    extendsWith: ["sanctions-check", "whois-lookup", "landing-page-roast"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company: {
+          type: "string",
+          description: "Company name, CIK, or ticker",
+        },
+        url: {
+          type: "string",
+          description: "Company website URL",
+        },
+      },
+      required: ["company", "url"],
+    },
+    exampleInput: { company: "Stripe", url: "https://stripe.com" },
+    exampleOutput: {
+      company_name: "Stripe, Inc.",
+      cik: "0001779474",
+      tech_stack: { frontend: "React", hosting: "AWS" },
+      seo_score: 88,
+      reputation_score: 95,
+    },
+    steps: [
+      {
+        capabilitySlug: "us-company-data",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company: "$input.company" },
+      },
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "seo-audit",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.url" },
+      },
+    ],
+  },
+
+  // ── 20b. Company Intelligence for SDR ──
+  {
+    slug: "company-intelligence-sdr",
+    name: "Company Intelligence for SDR",
+    marketingName: "Company Intelligence for SDR Agents",
+    description:
+      "Everything your SDR agent needs to decide whether to pursue a company: filing events, news sentiment, officers, tech stack, email patterns, social presence, domain age, and hiring signals. One call, 9 steps.",
+    longDescription:
+      "The complete company research package for AI SDR agents. Given a company name and domain, returns: recent filing events (SEC/Companies House), news sentiment from global media, company officers and directors from public records, technology stack detection, email pattern discovery, social media presence across 11 platforms, domain registration age and registrar (WHOIS), and hiring activity from job boards. Replaces the manual process of querying 9+ APIs separately. Every data point is quality-scored and sourced.",
+    agentDescription:
+      "research this company for outreach, SDR company intelligence, should I pursue this company, company research for sales, prospect company analysis, pre-outreach research, is this company worth targeting",
+    category: "sales-outreach",
+    priceCents: 250,
+    componentSumCents: 200,
+    valueTier: "verification",
+    maintenanceLevel: "low",
+    geography: "us-uk-eu",
+    targetAudience:
+      "Developers building AI SDR agents, outbound sales automation, or account research tools",
+    transparencyTag: "mixed",
+    extendsWith: ["sanctions-check", "vat-validate", "beneficial-ownership-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company_name: {
+          type: "string",
+          description: "Company name (e.g. Stripe, BMW, Novo Nordisk)",
+        },
+        domain: {
+          type: "string",
+          description: "Company website domain (e.g. stripe.com)",
+        },
+        country: {
+          type: "string",
+          description: "Country code (US, GB, DE, etc.) — improves accuracy for registry lookups",
+        },
+      },
+      required: ["company_name"],
+    },
+    exampleInput: { company_name: "Stripe", domain: "stripe.com", country: "US" },
+    exampleOutput: {
+      company_name: "Stripe, Inc.",
+      registry: { status: "active", cik: "0001779474" },
+      filing_events: [{ date: "2026-03-15", type: "financial_results" }],
+      news: { articles: 5, sentiment: "positive" },
+      tech_stack: ["Next.js", "Nginx", "Google Analytics"],
+      officers: [{ name: "John Collison", role: "President" }],
+      email_pattern: "first.last@stripe.com",
+    },
+    steps: [
+      // Parallel group 1: Company identity + filing events
+      {
+        capabilitySlug: "sec-filing-events",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company_name: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "company-news",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company_name: "$input.company_name", timespan: "14d" },
+      },
+      {
+        capabilitySlug: "officer-search",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company_name: "$input.company_name", country: "$input.country" },
+      },
+      // Parallel group 2: Domain-based signals (need domain input)
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "email-pattern-discover",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 6,
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { domain: "$input.domain" },
+      },
+      // Parallel group 3: Hiring + social + domain age
+      {
+        capabilitySlug: "job-board-search",
+        stepOrder: 7,
+        canParallel: true,
+        parallelGroup: 3,
+        inputMap: { query: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "social-profile-check",
+        stepOrder: 8,
+        canParallel: true,
+        parallelGroup: 3,
+        inputMap: { username: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "whois-lookup",
+        stepOrder: 9,
+        canParallel: true,
+        parallelGroup: 3,
+        inputMap: { domain: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── 20. Domain Trust Check ──
+  {
+    slug: "domain-trust",
+    name: "Domain Trust Check",
+    marketingName: "Domain Trust Check",
+    description:
+      "Given a domain, answer 'is this domain trustworthy?' — registration age, DNS, SSL, reputation score, and HTTP security headers.",
+    longDescription:
+      "Checks five trust signals for any domain: WHOIS registration age and registrar, DNS configuration health, SSL certificate validity, reputation score from threat intelligence feeds, and HTTP security header grade. Use for phishing detection, link safety checks, brand protection, or any time your agent needs to evaluate whether a domain is trustworthy before interacting with it.",
+    agentDescription:
+      "is this domain trustworthy, check domain trust, domain reputation check, is this website safe to visit, phishing check, link safety verification, should I trust this domain",
+    category: "security-risk",
+    priceCents: 40,
+    componentSumCents: 29,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "Anti-fraud teams, brand protection, phishing detection, or cybersecurity agents",
+    transparencyTag: "algorithmic",
+    extendsWith: ["tech-stack-detect", "backlink-check", "page-speed-test"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        domain: {
+          type: "string",
+          description: "Domain to check",
+        },
+      },
+      required: ["domain"],
+    },
+    exampleInput: { domain: "example.com" },
+    exampleOutput: {
+      registrar: "IANA",
+      registration_age_days: 10950,
+      dns_records: true,
+      ssl_valid: true,
+      reputation_score: 85,
+      header_security_grade: "C",
+    },
+    steps: [
+      {
+        capabilitySlug: "whois-lookup",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "dns-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "ssl-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$input.domain" },
+      },
+      {
+        capabilitySlug: "header-security-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.domain" },
+      },
+    ],
+  },
+
+  // ── Cross-model demand solutions (Batch 2 expansion) ──────────────────────
+
+  // ── Enhanced Due Diligence ──
+  {
+    slug: "enhanced-due-diligence",
+    name: "Enhanced Due Diligence",
+    marketingName: "Enhanced Due Diligence",
+    description:
+      "Comprehensive due diligence on a company: registry verification, beneficial ownership, sanctions screening, PEP screening, and adverse media analysis. Returns everything a compliance officer needs in one call.",
+    longDescription:
+      "Runs a full Enhanced Due Diligence (EDD) workflow: fetches official company registry data, looks up beneficial owners via Companies House for UK companies, screens against international sanctions lists, checks for Politically Exposed Persons, and searches for adverse media coverage. All steps run in parallel where possible for fast results.",
+    agentDescription:
+      "full company due diligence, EDD check, enhanced due diligence company, KYC AML company check, compliance screening company, beneficial ownership sanctions PEP adverse media",
+    category: "compliance",
+    priceCents: 300,
+    componentSumCents: 150,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "eu-global",
+    targetAudience: "Compliance teams, legal departments, investment firms, banks — anyone doing KYC/AML on companies",
+    transparencyTag: "mixed",
+    extendsWith: ["aml-risk-score", "insolvency-check", "domain-age-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company_name: { type: "string", description: "Company name to investigate" },
+        country_code: { type: "string", description: "ISO 2-letter country code" },
+        company_number: { type: "string", description: "Company registration number (optional)" },
+      },
+      required: ["company_name", "country_code"],
+    },
+    exampleInput: { company_name: "Tesco PLC", country_code: "GB", company_number: "00445790" },
+    exampleOutput: {
+      company: { name: "TESCO PLC", status: "active" },
+      beneficial_owners: [{ name: "Example Owner", percentage: 15.2 }],
+      sanctions: { match: false },
+      pep: { match: false },
+      adverse_media: { risk_level: "low", findings: [] },
+    },
+    steps: [
+      {
+        capabilitySlug: "uk-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { company_number: "$input.company_number" },
+      },
+      {
+        capabilitySlug: "beneficial-ownership-lookup",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { company_name: "$input.company_name", jurisdiction: "$input.country_code" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.company_name" },
+      },
+    ],
+  },
+
+  // ── Customer Risk Screen ──
+  {
+    slug: "customer-risk-screen",
+    name: "Customer Risk Screen",
+    marketingName: "Customer Risk Screen",
+    description:
+      "AML screening for customer onboarding: sanctions check, PEP screening, and adverse media analysis in one call. The essential compliance screen for every new customer.",
+    longDescription:
+      "Runs the three core AML screening steps in parallel: sanctions list check (EU, US OFAC, UN, UK), Politically Exposed Persons screening, and adverse media search. Designed for high-volume customer onboarding — all steps run in parallel for sub-second results on cached data.",
+    agentDescription:
+      "AML screening customer, KYC check person, customer onboarding compliance, sanctions PEP adverse media screen, new customer risk check",
+    category: "compliance",
+    priceCents: 100,
+    componentSumCents: 45,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "Any business onboarding customers — banks, fintechs, marketplaces, platforms with KYC requirements",
+    transparencyTag: "mixed",
+    extendsWith: ["aml-risk-score", "id-number-validate"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        name: { type: "string", description: "Person or company name to screen" },
+        country_code: { type: "string", description: "Country code (optional)" },
+        birth_date: { type: "string", description: "Date of birth for person screening (optional)" },
+      },
+      required: ["name"],
+    },
+    exampleInput: { name: "Jane Doe", country_code: "GB" },
+    exampleOutput: {
+      sanctions: { match: false },
+      pep: { is_pep: false, total_matches: 0 },
+      adverse_media: { risk_level: "none", total_findings: 0 },
+    },
+    steps: [
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.name", country: "$input.country_code", birth_date: "$input.birth_date" },
+      },
+      {
+        capabilitySlug: "adverse-media-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.name" },
+      },
+    ],
+  },
+
+  // ── Contact Verification ──
+  {
+    slug: "contact-verify",
+    name: "Contact Verification",
+    marketingName: "Contact Verification",
+    description:
+      "Verify a contact's email, phone number, and address in parallel. Returns validation status for each channel. Essential for CRM hygiene and lead qualification.",
+    longDescription:
+      "Validates all three contact channels simultaneously: email (format + MX records + disposable detection), phone (format + type classification + carrier), and address (geocoding + component extraction). All steps run in parallel — results in under 2 seconds.",
+    agentDescription:
+      "verify contact details, validate email phone address, CRM data hygiene, lead qualification check, contact data verification",
+    category: "sales-outreach",
+    priceCents: 25,
+    componentSumCents: 8,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience: "Sales teams, CRM administrators, marketing ops — anyone maintaining contact databases",
+    transparencyTag: "mixed",
+    extendsWith: ["phone-type-detect", "email-reputation-score", "address-geocode"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        email: { type: "string", description: "Email address to verify" },
+        phone_number: { type: "string", description: "Phone number to verify (optional)" },
+        address: { type: "string", description: "Postal address to verify (optional)" },
+      },
+      required: ["email"],
+    },
+    exampleInput: { email: "jane@acme.com", phone_number: "+447911123456", address: "10 Downing Street, London, UK" },
+    exampleOutput: {
+      email: { valid: true, has_mx: true },
+      phone: { valid: true, type: "mobile" },
+      address: { valid: true, formatted: "10 Downing Street, London SW1A 2AA, UK" },
+    },
+    steps: [
+      {
+        capabilitySlug: "email-validate",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { email: "$input.email" },
+      },
+      {
+        capabilitySlug: "phone-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { phone_number: "$input.phone_number" },
+      },
+      {
+        capabilitySlug: "address-validate",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { address: "$input.address" },
+      },
+    ],
+  },
+
+  // ── Invoice Processing Pipeline ──
+  {
+    slug: "invoice-process",
+    name: "Invoice Processing Pipeline",
+    marketingName: "Invoice Processing Pipeline",
+    description:
+      "End-to-end invoice verification: extract structured data from the invoice, validate the vendor's VAT number, verify the payment IBAN, and convert currency if needed. Returns everything an AP automation agent needs.",
+    longDescription:
+      "Processes an invoice in four steps: (1) extracts structured data (vendor, amounts, line items, VAT, IBAN) using AI, (2) validates the VAT number against VIES, (3) validates the IBAN checksum and bank, (4) converts the total to a target currency. Steps 2-3 run in parallel after extraction.",
+    agentDescription:
+      "process invoice, extract invoice data, verify invoice VAT IBAN, AP automation, accounts payable processing, invoice validation pipeline",
+    category: "data-research",
+    priceCents: 50,
+    componentSumCents: 23,
+    valueTier: "verification",
+    maintenanceLevel: "low",
+    geography: "eu-global",
+    targetAudience: "Finance teams, AP automation, ERP integrations — anyone processing invoices",
+    transparencyTag: "mixed",
+    extendsWith: ["invoice-validate", "payment-reference-generate"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        text: { type: "string", description: "Invoice text or OCR output" },
+        target_currency: { type: "string", description: "Convert amounts to this currency (optional)" },
+      },
+      required: ["text"],
+    },
+    exampleInput: {
+      text: "Invoice #2024-001\nFrom: Acme Ltd, VAT: GB123456789\nIBAN: GB82WEST12345698765432\nTotal: €1,250.00",
+      target_currency: "SEK",
+    },
+    exampleOutput: {
+      extracted: { vendor: "Acme Ltd", vat_number: "GB123456789", total: 1250.0 },
+      vat_valid: true,
+      iban_valid: true,
+      converted: { amount: 14375.0, currency: "SEK" },
+    },
+    steps: [
+      {
+        capabilitySlug: "invoice-extract",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { text: "$input.text" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$steps[0].vat_number" },
+      },
+      {
+        capabilitySlug: "iban-validate",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { iban: "$steps[0].iban" },
+      },
+      {
+        capabilitySlug: "currency-convert",
+        stepOrder: 4,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { from: "$steps[0].currency", to: "$input.target_currency", amount: "$steps[0].total_amount" },
+      },
+    ],
+  },
+
+  // ── Vendor Onboarding Check ──
+  {
+    slug: "vendor-onboard",
+    name: "Vendor Onboarding Check",
+    marketingName: "Vendor Onboarding Check",
+    description:
+      "Everything a procurement agent needs to onboard a new vendor: company verification, VAT validation, bank account verification, sanctions screening, and address validation. One call replaces a 45-minute manual check.",
+    longDescription:
+      "Runs five verification steps: (1) fetches company registry data, (2-5) validates VAT number, verifies IBAN, screens against sanctions lists, and validates the vendor address — all in parallel after the initial company lookup.",
+    agentDescription:
+      "vendor onboarding check, onboard new supplier, procurement compliance, verify vendor company VAT IBAN sanctions, supplier due diligence",
+    category: "compliance",
+    priceCents: 150,
+    componentSumCents: 104,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "eu-global",
+    targetAudience: "Procurement teams, vendor management, supply chain compliance",
+    transparencyTag: "mixed",
+    extendsWith: ["beneficial-ownership-lookup", "insolvency-check", "domain-age-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        company_name: { type: "string", description: "Vendor company name" },
+        country_code: { type: "string", description: "Vendor country code" },
+        vat_number: { type: "string", description: "Vendor VAT number (optional)" },
+        iban: { type: "string", description: "Vendor bank IBAN (optional)" },
+        address: { type: "string", description: "Vendor address (optional)" },
+      },
+      required: ["company_name", "country_code"],
+    },
+    exampleInput: { company_name: "Volvo AB", country_code: "SE", vat_number: "SE556012573201", iban: "SE3550000000054910000003" },
+    exampleOutput: {
+      company: { name: "Volvo AB", status: "active" },
+      vat_valid: true,
+      iban_valid: true,
+      sanctions_match: false,
+      address_valid: true,
+    },
+    steps: [
+      {
+        capabilitySlug: "swedish-company-data",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { org_number: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "vat-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { vat_number: "$input.vat_number" },
+      },
+      {
+        capabilitySlug: "iban-validate",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { iban: "$input.iban" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.company_name" },
+      },
+      {
+        capabilitySlug: "address-validate",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { address: "$input.address" },
+      },
+    ],
+  },
+
+  // ── HR Candidate Screening ──
+  {
+    slug: "hr-candidate-screen",
+    name: "HR Candidate Screening",
+    marketingName: "HR Candidate Screening",
+    description:
+      "Screen a job candidate: parse their resume, validate their email and phone, and run sanctions and PEP checks. Returns structured candidate data with verification status.",
+    longDescription:
+      "Parses the resume to extract candidate name and details, then runs four verification steps in parallel: email validation, phone validation, sanctions screening, and PEP check. Designed for HR tech platforms and recruitment agencies in regulated industries.",
+    agentDescription:
+      "screen job candidate, HR background check, recruitment compliance, verify candidate email phone sanctions PEP, candidate screening pipeline",
+    category: "data-research",
+    priceCents: 80,
+    componentSumCents: 30,
+    valueTier: "verification",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "HR tech, recruitment platforms, staffing agencies, compliance-heavy industries",
+    transparencyTag: "mixed",
+    extendsWith: ["id-number-validate", "adverse-media-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        resume_text: { type: "string", description: "Resume/CV text content" },
+        email: { type: "string", description: "Candidate email (optional)" },
+        phone_number: { type: "string", description: "Candidate phone (optional)" },
+      },
+      required: ["resume_text"],
+    },
+    exampleInput: {
+      resume_text: "Jane Smith\nSenior Engineer at Acme Corp\njane@acme.com\n+44 7911 123456",
+      email: "jane@acme.com",
+      phone_number: "+447911123456",
+    },
+    exampleOutput: {
+      candidate: { name: "Jane Smith", skills: ["engineering"] },
+      email_valid: true,
+      phone_valid: true,
+      sanctions_match: false,
+      pep_match: false,
+    },
+    steps: [
+      {
+        capabilitySlug: "resume-parse",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { text: "$input.resume_text" },
+      },
+      {
+        capabilitySlug: "email-validate",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { email: "$input.email" },
+      },
+      {
+        capabilitySlug: "phone-validate",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { phone_number: "$input.phone_number" },
+      },
+      {
+        capabilitySlug: "sanctions-check",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].name" },
+      },
+      {
+        capabilitySlug: "pep-check",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$steps[0].name" },
+      },
+    ],
+  },
+
+  // ── Crypto/Web3 solutions ──────────────────────────────────────────────────
+
+  {
+    slug: "token-project-dd",
+    name: "Token Project Due Diligence",
+    marketingName: "Token Project Due Diligence",
+    description:
+      "Your agent is evaluating a token project or crypto startup. This solution checks the team's web presence, verifies the company entity, screens for sanctions, and assesses domain trust — the off-chain due diligence that on-chain analysis can't provide.",
+    longDescription:
+      "Detects the project's tech stack, validates SSL certificates, assesses domain reputation, looks up WHOIS records, and screens the team entity against global sanctions lists. Covers the off-chain risk factors that blockchain analytics miss: is the website real, is the domain trustworthy, is the team sanctioned?",
+    agentDescription:
+      "crypto due diligence, token project check, verify crypto startup, web3 team verification, off-chain risk check, token launch evaluation, crypto VC due diligence, defi project assessment",
+    category: "data-research",
+    priceCents: 200,
+    componentSumCents: 63,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "Crypto VCs, token analysts, DeFi protocol evaluators, ai16z-style investment agents",
+    transparencyTag: "mixed",
+    extendsWith: ["landing-page-roast", "social-profile-check", "backlink-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        project_url: { type: "string", format: "uri", description: "Token project website URL" },
+        team_entity_name: { type: "string", description: "Legal entity or team name behind the project" },
+      },
+      required: ["project_url", "team_entity_name"],
+    },
+    exampleInput: { project_url: "https://uniswap.org", team_entity_name: "Uniswap Labs" },
+    exampleOutput: {
+      tech_stack: { frontend: "React", hosting: "Vercel" },
+      ssl_valid: true,
+      domain_reputation: "clean",
+      domain_age_days: 1825,
+      is_sanctioned: false,
+    },
+    steps: [
+      { capabilitySlug: "tech-stack-detect", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.project_url" } },
+      { capabilitySlug: "ssl-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.project_url" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.project_url" } },
+      { capabilitySlug: "whois-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.project_url" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.team_entity_name" } },
+    ],
+  },
+
+  {
+    slug: "defi-risk-check",
+    name: "DeFi Protocol Risk Check",
+    marketingName: "DeFi Protocol Risk Check",
+    description:
+      "Before your agent deposits funds into a DeFi protocol, verify the team isn't sanctioned, the domain is legitimate, and the SSL certificate is valid. The off-chain risk layer for on-chain DeFi.",
+    longDescription:
+      "Validates the protocol's SSL certificate, assesses domain reputation, screens the protocol name against global sanctions lists, and looks up WHOIS records. Provides the off-chain trust signals that complement on-chain protocol analysis.",
+    agentDescription:
+      "defi risk check, protocol safety check, is this defi protocol safe, verify defi protocol, yield farming risk, defi due diligence, smart contract risk, protocol trust score",
+    category: "compliance-verification",
+    priceCents: 150,
+    componentSumCents: 43,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "DeFi aggregators, yield farming agents, portfolio managers, risk teams",
+    transparencyTag: "mixed",
+    extendsWith: ["privacy-policy-analyze", "cookie-scan"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        protocol_url: { type: "string", format: "uri", description: "DeFi protocol website URL" },
+        protocol_name: { type: "string", description: "Protocol name (e.g. Aave, Compound)" },
+      },
+      required: ["protocol_url", "protocol_name"],
+    },
+    exampleInput: { protocol_url: "https://aave.com", protocol_name: "Aave" },
+    exampleOutput: {
+      ssl_valid: true,
+      domain_reputation: "clean",
+      is_sanctioned: false,
+      domain_registrar: "MarkMonitor",
+    },
+    steps: [
+      { capabilitySlug: "ssl-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.protocol_url" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.protocol_url" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { name: "$input.protocol_name" } },
+      { capabilitySlug: "whois-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.protocol_url" } },
+    ],
+  },
+
+  {
+    slug: "crypto-counterparty-kyb",
+    name: "Crypto Counterparty KYB",
+    marketingName: "Crypto Counterparty KYB",
+    description:
+      "Know Your Business for crypto counterparties. Verify the legal entity exists, check sanctions lists, validate their web presence, and assess domain trust. The compliance layer agents need before transacting with a new counterparty.",
+    longDescription:
+      "Screens the counterparty against global sanctions lists (OFAC, EU, UN), assesses domain reputation, validates SSL certificates, looks up WHOIS records, and checks email deliverability for the entity's domain. Covers the off-chain compliance requirements for crypto-native businesses.",
+    agentDescription:
+      "crypto KYB, know your business crypto, verify crypto counterparty, OTC desk compliance, exchange counterparty check, institutional defi compliance, crypto entity verification",
+    category: "compliance-verification",
+    priceCents: 180,
+    componentSumCents: 53,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "OTC desks, crypto exchanges, institutional DeFi, compliance teams",
+    transparencyTag: "mixed",
+    extendsWith: ["company-enrich", "pep-check", "adverse-media-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        entity_name: { type: "string", description: "Legal entity or company name" },
+        entity_website: { type: "string", format: "uri", description: "Entity's website URL" },
+      },
+      required: ["entity_name", "entity_website"],
+    },
+    exampleInput: { entity_name: "Circle Internet Financial", entity_website: "https://circle.com" },
+    exampleOutput: {
+      is_sanctioned: false,
+      domain_reputation: "clean",
+      ssl_valid: true,
+      email_deliverable: true,
+      domain_registrar: "MarkMonitor",
+    },
+    steps: [
+      { capabilitySlug: "sanctions-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.entity_website" } },
+      { capabilitySlug: "ssl-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.entity_website" } },
+      { capabilitySlug: "whois-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.entity_website" } },
+      { capabilitySlug: "email-deliverability-check", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.entity_website" } },
+    ],
+  },
+
+  // ── Web3 Solutions ────────────────────────────────────────────────────────
+
+  {
+    slug: "token-project-dd",
+    name: "Token Project Due Diligence",
+    marketingName: "Token Project Due Diligence",
+    description:
+      "Evaluates a token project's legitimacy before investment or integration. Checks domain infrastructure, security posture, sanctions status, and technology stack. The answer to \"is this token project real?\"",
+    category: "security-risk",
+    priceCents: 75,
+    componentSumCents: 52,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "Web3 fund compliance agents, DeFi risk bots, ElizaOS agents, Solana Agent Kit users, x402 clients evaluating counterparties",
+    transparencyTag: "mixed",
+    extendsWith: ["crypto-price", "us-company-data"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        domain: {
+          type: "string",
+          description: "Token project website domain (e.g. uniswap.org)",
+        },
+        entity_name: {
+          type: "string",
+          description: "Project or company name for sanctions screening",
+        },
+      },
+      required: ["domain", "entity_name"],
+    },
+    exampleInput: { domain: "uniswap.org", entity_name: "Uniswap Labs" },
+    exampleOutput: {
+      registrar: "Cloudflare Inc.",
+      ssl_valid: true,
+      days_until_expiry: 312,
+      reputation_score: 91,
+      header_security_grade: "B",
+      is_sanctioned: false,
+      tech_stack: { frontend: "React", hosting: "Cloudflare" },
+    },
+    steps: [
+      { capabilitySlug: "whois-lookup", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "ssl-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "dns-lookup", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "header-security-check", stepOrder: 5, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.domain" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 6, canParallel: true, parallelGroup: 1, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "tech-stack-detect", stepOrder: 7, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.domain" } },
+    ],
+  },
+  {
+    slug: "defi-protocol-risk",
+    name: "DeFi Protocol Risk Check",
+    marketingName: "DeFi Protocol Risk Check",
+    description:
+      "Evaluates a DeFi protocol's infrastructure security before interaction. Checks SSL certificate health, HTTP security headers, DNS configuration, domain reputation, and technology surface area. Answers \"is this protocol's infrastructure secure?\"",
+    category: "security-risk",
+    priceCents: 50,
+    componentSumCents: 32,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "DeFi risk agents, yield farming bots, portfolio management agents, x402 clients, Solana Agent Kit users",
+    transparencyTag: "mixed",
+    extendsWith: ["crypto-price", "whois-lookup", "sanctions-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        domain: {
+          type: "string",
+          description: "DeFi protocol domain (e.g. app.aave.com)",
+        },
+      },
+      required: ["domain"],
+    },
+    exampleInput: { domain: "app.aave.com" },
+    exampleOutput: {
+      ssl_valid: true,
+      days_until_expiry: 245,
+      header_security_grade: "A",
+      reputation_score: 94,
+      dns_records: true,
+      tech_stack: { frontend: "React", hosting: "Cloudflare" },
+    },
+    steps: [
+      { capabilitySlug: "ssl-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "header-security-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.domain" } },
+      { capabilitySlug: "dns-lookup", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "tech-stack-detect", stepOrder: 5, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.domain" } },
+    ],
+  },
+  {
+    slug: "web3-counterparty-kyb",
+    name: "Web3 Counterparty KYB",
+    marketingName: "Web3 Counterparty KYB",
+    description:
+      "Counterparty due diligence for Web3 companies, DAOs, and token projects. Sanctions screening on the entity plus full domain infrastructure audit. Answers \"should we transact with this counterparty?\"",
+    category: "security-risk",
+    priceCents: 70,
+    componentSumCents: 49,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "Web3 OTC desks, DAO treasury agents, DeFi lending protocol risk engines, x402 payment agents, ElizaOS agents",
+    transparencyTag: "mixed",
+    extendsWith: ["crypto-price", "us-company-data", "tech-stack-detect"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        entity_name: {
+          type: "string",
+          description: "Company, DAO, or project name for sanctions screening",
+        },
+        domain: {
+          type: "string",
+          description: "Counterparty website domain",
+        },
+      },
+      required: ["entity_name", "domain"],
+    },
+    exampleInput: { entity_name: "Circle Internet Financial", domain: "circle.com" },
+    exampleOutput: {
+      is_sanctioned: false,
+      confidence: "high",
+      registrar: "MarkMonitor Inc.",
+      ssl_valid: true,
+      reputation_score: 96,
+      header_security_grade: "A",
+      dns_records: true,
+    },
+    steps: [
+      { capabilitySlug: "sanctions-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "whois-lookup", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "ssl-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+      { capabilitySlug: "header-security-check", stepOrder: 5, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.domain" } },
+      { capabilitySlug: "dns-lookup", stepOrder: 6, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.domain" } },
+    ],
+  },
+
+  // ═══════════════════════════════════════════════════════════════════════════
+  // Web3 Agent Solutions — enriched multi-source workflows
+  // ═══════════════════════════════════════════════════════════════════════════
+
+  {
+    slug: "web3-counterparty-dd",
+    name: "Web3 Counterparty Due Diligence",
+    marketingName: "Web3 Counterparty Due Diligence",
+    description: "Full compliance check on a wallet address before transacting. Combines on-chain risk scoring with off-chain entity compliance — wallet risk, wallet age, ENS identity resolution, sanctions screening, PEP check, and adverse media scan.",
+    agentDescription: "web3 counterparty due diligence, wallet kyc, wallet compliance check, defi counterparty risk, on-chain aml, wallet sanctions check, crypto counterparty screening",
+    category: "web3",
+    priceCents: 15,
+    componentSumCents: 12,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "DeFi lending agents, OTC desks, DAO treasury managers, compliance-aware trading bots",
+    transparencyTag: "mixed",
+    extendsWith: ["beneficial-ownership-lookup", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        wallet_address: { type: "string", description: "Wallet address to investigate (0x...)" },
+        entity_name: { type: "string", description: "Entity or person name for sanctions/PEP check (if known)" },
+        chain_id: { type: "string", description: "Chain ID, default 1 (Ethereum)" },
+      },
+      required: ["wallet_address"],
+    },
+    exampleInput: { wallet_address: "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045", entity_name: "Vitalik Buterin", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "wallet-risk-score", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "wallet-age-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "ens-reverse-lookup", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "pep-check", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "adverse-media-check", stepOrder: 6, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.entity_name" } },
+    ],
+  },
+  {
+    slug: "web3-token-safety",
+    name: "Web3 Token Safety Check",
+    marketingName: "Web3 Token Safety Check",
+    description: "Before buying a token, verify the contract isn't a scam. Checks for honeypot behavior, hidden ownership, mint functions, and verifies contract source is public. Also checks the deployer wallet for fraud history and the project's domain reputation.",
+    agentDescription: "token safety check, is this token a scam, honeypot check, rug pull detection, token contract audit, defi token risk, erc20 safety",
+    category: "web3",
+    priceCents: 8,
+    componentSumCents: 8,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience: "DeFi trading agents, portfolio managers, token listing screeners, investor research bots",
+    transparencyTag: null,
+    extendsWith: ["whois-lookup", "ssl-check", "tech-stack-detect"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        contract_address: { type: "string", description: "Token contract address (0x...)" },
+        project_domain: { type: "string", description: "Project website domain for reputation check" },
+        chain_id: { type: "string", description: "Chain ID, default 1" },
+      },
+      required: ["contract_address"],
+    },
+    exampleInput: { contract_address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48", project_domain: "circle.com", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "token-security-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { contract_address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "contract-verify-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { contract_address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "wallet-risk-score", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.project_domain" } },
+    ],
+  },
+  {
+    slug: "web3-pre-tx-gate",
+    name: "Web3 Pre-Transaction Risk Gate",
+    marketingName: "Web3 Pre-Transaction Risk Gate",
+    description: "Should your agent execute this transaction? Single-call go/no-go middleware for DeFi agents. Checks wallet risk, token safety, approval security, current gas costs, and sanctions status. Returns a risk verdict.",
+    agentDescription: "pre-transaction check, should i execute this trade, defi safety gate, transaction risk check, agent middleware, trading bot safety, smart contract interaction check",
+    category: "web3",
+    priceCents: 15,
+    componentSumCents: 10,
+    valueTier: "compliance",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience: "Autonomous DeFi agents, trading bots with safety rails, agent frameworks needing pre-flight checks",
+    transparencyTag: null,
+    extendsWith: ["phishing-site-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        wallet_address: { type: "string", description: "Counterparty or target wallet address" },
+        contract_address: { type: "string", description: "Token contract if relevant" },
+        entity_name: { type: "string", description: "Entity name for sanctions check" },
+        chain_id: { type: "string", description: "Chain ID, default 1" },
+      },
+      required: ["wallet_address"],
+    },
+    exampleInput: { wallet_address: "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045", contract_address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48", entity_name: "Circle", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "wallet-risk-score", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "token-security-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { contract_address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "approval-security-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "gas-price-check", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { chain_id: "$input.chain_id" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.entity_name" } },
+    ],
+  },
+  {
+    slug: "web3-dapp-trust",
+    name: "Web3 dApp Trust Scan",
+    marketingName: "Web3 dApp Trust Scan",
+    description: "Before connecting your wallet to a dApp, verify it's not a phishing clone. Combines GoPlus phishing detection with deep domain intelligence — reputation scoring, SSL verification, WHOIS data, and technology stack analysis.",
+    agentDescription: "dapp trust check, is this dapp safe, phishing detection, wallet connect safety, dapp security scan, web3 phishing check, fake dapp detection",
+    category: "web3",
+    priceCents: 8,
+    componentSumCents: 8,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience: "Wallet security agents, browser extension bots, dApp aggregator screeners",
+    transparencyTag: null,
+    extendsWith: ["header-security-check", "dns-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        url: { type: "string", description: "dApp URL to check" },
+      },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://app.uniswap.org" },
+    steps: [
+      { capabilitySlug: "phishing-site-check", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { url: "$input.url" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.url" } },
+      { capabilitySlug: "ssl-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { domain: "$input.url" } },
+      { capabilitySlug: "whois-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.url" } },
+      { capabilitySlug: "tech-stack-detect", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { url: "$input.url" } },
+    ],
+  },
+  {
+    slug: "web3-protocol-health",
+    name: "Web3 Protocol Health Report",
+    marketingName: "Web3 Protocol Health Report",
+    description: "Full health snapshot of any DeFi protocol. TVL, fee revenue, stablecoin exposure, plus infrastructure trust assessment. Answers: is this protocol real AND healthy?",
+    agentDescription: "protocol health check, defi protocol analysis, tvl check, protocol revenue, is this protocol healthy, defi due diligence, yield farming research",
+    category: "web3",
+    priceCents: 8,
+    componentSumCents: 8,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "DeFi yield agents, protocol comparison tools, treasury allocation bots, risk dashboards",
+    transparencyTag: null,
+    extendsWith: ["landing-page-roast", "social-profile-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        protocol: { type: "string", description: "Protocol name or DeFi Llama slug (e.g., 'aave-v3', 'uniswap')" },
+        protocol_domain: { type: "string", description: "Protocol website domain for infrastructure check" },
+      },
+      required: ["protocol"],
+    },
+    exampleInput: { protocol: "aave-v3", protocol_domain: "aave.com" },
+    steps: [
+      { capabilitySlug: "protocol-tvl-lookup", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { protocol: "$input.protocol" } },
+      { capabilitySlug: "protocol-fees-lookup", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { protocol: "$input.protocol" } },
+      { capabilitySlug: "stablecoin-flow-check", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { chain: "Ethereum" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.protocol_domain" } },
+      { capabilitySlug: "ssl-check", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.protocol_domain" } },
+    ],
+  },
+  {
+    slug: "web3-pre-trade",
+    name: "Web3 Pre-Trade Intelligence",
+    marketingName: "Web3 Pre-Trade Intelligence",
+    description: "Everything a trading agent needs before placing a trade. Market price, token contract safety, protocol health, deployer wallet risk, market sentiment, and gas costs — all in one call.",
+    agentDescription: "pre-trade intelligence, should i buy this token, trading bot research, defi trade analysis, token price and safety, market sentiment check, gas cost check",
+    category: "web3",
+    priceCents: 12,
+    componentSumCents: 12,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "Autonomous trading bots, portfolio rebalancing agents, DeFi yield optimizers, quantitative trading systems",
+    transparencyTag: null,
+    extendsWith: ["stablecoin-flow-check", "wallet-age-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        token_id: { type: "string", description: "CoinGecko token ID or symbol (e.g., 'ethereum', 'bitcoin')" },
+        contract_address: { type: "string", description: "Token contract address for security check" },
+        protocol: { type: "string", description: "DeFi Llama protocol slug for TVL check" },
+        chain_id: { type: "string", description: "Chain ID, default 1" },
+      },
+      required: ["token_id"],
+    },
+    exampleInput: { token_id: "ethereum", contract_address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48", protocol: "aave-v3", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "crypto-price", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { coin: "$input.token_id" } },
+      { capabilitySlug: "token-security-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { contract_address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "protocol-tvl-lookup", stepOrder: 3, canParallel: true, parallelGroup: 1, inputMap: { protocol: "$input.protocol" } },
+      { capabilitySlug: "wallet-risk-score", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { address: "$input.contract_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "fear-greed-index", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: {} },
+      { capabilitySlug: "gas-price-check", stepOrder: 6, canParallel: true, parallelGroup: 2, inputMap: { chain_id: "$input.chain_id" } },
+    ],
+  },
+  {
+    slug: "web3-wallet-snapshot",
+    name: "Web3 Wallet Portfolio Snapshot",
+    marketingName: "Web3 Wallet Portfolio Snapshot",
+    description: "Understand any wallet before acting. Balance, recent transactions, age, ENS identity, and current market prices — the quickest way to size up a wallet.",
+    agentDescription: "wallet snapshot, wallet portfolio, wallet analysis, wallet balance check, wallet history, wallet identity, wallet age check",
+    category: "web3",
+    priceCents: 5,
+    componentSumCents: 5,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "Portfolio tracking agents, wallet analytics tools, counterparty screening, airdrop eligibility checkers",
+    transparencyTag: null,
+    extendsWith: ["wallet-risk-score", "approval-security-check"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        wallet_address: { type: "string", description: "Wallet address (0x...)" },
+        chain_id: { type: "string", description: "Chain ID, default 1" },
+      },
+      required: ["wallet_address"],
+    },
+    exampleInput: { wallet_address: "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "wallet-balance-lookup", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "wallet-transactions-lookup", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id", limit: "10" } },
+      { capabilitySlug: "wallet-age-check", stepOrder: 3, canParallel: true, parallelGroup: 2, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "ens-reverse-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { address: "$input.wallet_address" } },
+      { capabilitySlug: "crypto-price", stepOrder: 5, canParallel: true, parallelGroup: 2, inputMap: { coin: "ethereum" } },
+    ],
+  },
+  {
+    slug: "web3-vasp-check",
+    name: "Web3 VASP Compliance Check",
+    marketingName: "Web3 VASP Compliance Check",
+    description: "Before interacting with a crypto service provider, verify they're licensed under EU MiCA regulation. Checks the ESMA register of authorized CASPs, the non-compliant entity list, sanctions status, and domain reputation.",
+    agentDescription: "vasp compliance, mica check, casp verification, crypto exchange license check, eu crypto regulation, vasp authorization, exchange compliance",
+    category: "web3",
+    priceCents: 10,
+    componentSumCents: 8,
+    valueTier: "compliance",
+    maintenanceLevel: "low",
+    geography: "eu",
+    targetAudience: "Compliance agents, exchange screening bots, crypto fund due diligence, VASP onboarding automation",
+    transparencyTag: null,
+    extendsWith: ["pep-check", "adverse-media-check", "whois-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        entity_name: { type: "string", description: "CASP/VASP name or company name" },
+        website: { type: "string", description: "Entity website domain" },
+      },
+      required: ["entity_name"],
+    },
+    exampleInput: { entity_name: "Coinbase", website: "coinbase.com" },
+    steps: [
+      { capabilitySlug: "vasp-verify", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { entity_name: "$input.entity_name", website: "$input.website" } },
+      { capabilitySlug: "vasp-non-compliant-check", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { entity_name: "$input.entity_name", website: "$input.website" } },
+      { capabilitySlug: "sanctions-check", stepOrder: 3, canParallel: true, parallelGroup: 2, inputMap: { name: "$input.entity_name" } },
+      { capabilitySlug: "domain-reputation", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { domain: "$input.website" } },
+    ],
+  },
+  {
+    slug: "web3-wallet-identity",
+    name: "Web3 Wallet Identity Resolution",
+    marketingName: "Web3 Wallet Identity Resolution",
+    description: "Given a wallet address, discover everything knowable about the entity behind it. Starts with ENS reverse lookup to find a human-readable name, then assesses risk, age, and holdings. The entry point to the full compliance pipeline.",
+    agentDescription: "wallet identity, who owns this wallet, wallet deanonymization, wallet reputation, wallet kyc, ens lookup, wallet profiling",
+    category: "web3",
+    priceCents: 10,
+    componentSumCents: 8,
+    valueTier: "data-lookup",
+    maintenanceLevel: "low",
+    geography: "global",
+    targetAudience: "KYC/AML agents, fraud investigation bots, counterparty research tools, compliance automation",
+    transparencyTag: null,
+    extendsWith: ["sanctions-check", "pep-check", "company-enrich"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        wallet_address: { type: "string", description: "Wallet address to investigate (0x...)" },
+        chain_id: { type: "string", description: "Chain ID, default 1" },
+      },
+      required: ["wallet_address"],
+    },
+    exampleInput: { wallet_address: "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045", chain_id: "1" },
+    steps: [
+      { capabilitySlug: "ens-reverse-lookup", stepOrder: 1, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address" } },
+      { capabilitySlug: "wallet-risk-score", stepOrder: 2, canParallel: true, parallelGroup: 1, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "wallet-age-check", stepOrder: 3, canParallel: true, parallelGroup: 2, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+      { capabilitySlug: "wallet-balance-lookup", stepOrder: 4, canParallel: true, parallelGroup: 2, inputMap: { address: "$input.wallet_address", chain_id: "$input.chain_id" } },
+    ],
+  },
+
+  // ── Dependency Risk Check ──
+  {
+    slug: "nl-property-check",
+    name: "Netherlands Property Check",
+    marketingName: "Netherlands Property Intelligence",
+    description:
+      "Comprehensive Dutch property market intelligence. Combines BAG address data, WOZ tax assessments, energy labels, housing prices, and regional statistics into one call.",
+    longDescription:
+      "Full Netherlands property analysis: official BAG address/building data from Kadaster, municipality-level WOZ tax values and trends from CBS, energy performance labels from EP-Online, national housing price index, and detailed housing statistics. Covers construction year, floor area, ownership patterns, and market trends. Three CBS data sources require no API key; BAG and EP-Online use free registered keys.",
+    agentDescription:
+      "netherlands property check, dutch real estate data, BAG address lookup, WOZ waarde, energielabel, housing price index, CBS housing statistics, kadaster, nl property",
+    category: "data-lookup",
+    priceCents: 19,
+    componentSumCents: 19,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "eu",
+    targetAudience:
+      "Real estate agents, property investors, mortgage brokers, proptech companies operating in the Netherlands",
+    transparencyTag: null,
+    extendsWith: ["dutch-company-data"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        postcode: { type: "string", description: "Dutch postcode (format 1234AB)" },
+        huisnummer: { type: "string", description: "House number" },
+        city: { type: "string", description: "Municipality name (e.g. 'Amsterdam') — used for CBS stats" },
+      },
+      required: ["city"],
+    },
+    exampleInput: { postcode: "1012JS", huisnummer: "1", city: "Amsterdam" },
+    exampleOutput: {
+      bag_address: { street: "Dam", city: "Amsterdam", construction_year: 1900 },
+      woz_value: { average_woz_value_eur: 516000, latest_year: 2023 },
+      energy_label: { energy_label: "C" },
+      housing_prices: { price_index: 138.2, average_sale_price_eur: 438000 },
+      housing_stats: { sale_prices: { average_sale_price_eur: 520000 } },
+    },
+    steps: [
+      {
+        capabilitySlug: "nl-bag-address",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { postcode: "$input.postcode", huisnummer: "$input.huisnummer" },
+      },
+      {
+        capabilitySlug: "nl-woz-value",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { city: "$input.city" },
+      },
+      {
+        capabilitySlug: "nl-energy-label",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { postcode: "$input.postcode", huisnummer: "$input.huisnummer" },
+      },
+      {
+        capabilitySlug: "nl-housing-price-index",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { months: "12" },
+      },
+      {
+        capabilitySlug: "nl-housing-stats",
+        stepOrder: 5,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { city: "$input.city" },
+      },
+    ],
+  },
+  {
+    slug: "dependency-risk-check",
+    name: "Dependency Risk Check",
+    marketingName: "Dependency Risk Assessment",
+    description:
+      "Complete security and license risk assessment for any npm or PyPI package. Combines CVE scanning (OSV.dev), OpenSSF Scorecard analysis, freshness check, and license compatibility verification into one call.",
+    longDescription:
+      "Answers: is this dependency safe to use in my project? Runs a composite CVE check via OSV.dev, fetches the OpenSSF Scorecard, checks package freshness and deprecation status, and verifies license compatibility with your use case. All data sources are free and require no API keys.",
+    agentDescription:
+      "dependency risk check, package security audit, npm pypi vulnerability scan, license compatibility, OpenSSF scorecard, CVE check, dependency audit",
+    category: "security",
+    priceCents: 25,
+    componentSumCents: 20,
+    valueTier: "verification",
+    maintenanceLevel: "near-zero",
+    geography: "global",
+    targetAudience:
+      "Developers using coding agents (Claude Code, Cursor, Windsurf), CI/CD pipelines, security teams evaluating dependencies",
+    transparencyTag: null,
+    extendsWith: ["npm-package-info", "pypi-package-info", "cve-lookup"],
+    inputSchema: {
+      type: "object",
+      properties: {
+        name: { type: "string", description: "Package name (e.g. 'express', 'requests')" },
+        version: { type: "string", description: "Version to audit (optional — defaults to latest)" },
+        ecosystem: { type: "string", enum: ["npm", "pypi"], description: "Package ecosystem (optional — auto-detected)" },
+        use_case: { type: "string", enum: ["commercial", "open-source", "saas", "internal"], description: "How the software will be used (for license check). Default: commercial" },
+      },
+      required: ["name"],
+    },
+    exampleInput: { name: "express", version: "4.18.2", use_case: "commercial" },
+    exampleOutput: {
+      security_audit: { risk_score: 82, risk_level: "low", vulnerabilities: { total: 1 } },
+      license_check: { compatible: true, conflicts: [] },
+    },
+    steps: [
+      {
+        capabilitySlug: "package-security-audit",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { name: "$input.name", version: "$input.version", ecosystem: "$input.ecosystem" },
+      },
+      {
+        capabilitySlug: "license-compatibility-check",
+        stepOrder: 2,
+        // Was parallelGroup 1, racing the step it reads from. Same group runs
+        // concurrently, so `$steps[0].license.spdx` could resolve before
+        // package-security-audit had produced it — the executor's null
+        // fallback then silently dropped the licence list and the check ran
+        // on nothing. Moved to its own group so the dependency is honoured.
+        // Found 2026-08-16 by the execution-order test added with the growth
+        // bundles; this bundle has no external sales, so nothing regressed.
+        canParallel: true,
+        parallelGroup: 2,
+        inputMap: { licenses: "$steps[0].license.spdx", use_case: "$input.use_case" },
+      },
+    ],
+  },
+
+  // ─────────────────────────────────────────────────────────────────────────
+  // Growth bundles, added 2026-08-16.
+  //
+  // Modelled on `lead-email-verify`, which is the only bundle of the 98 we
+  // had built that customers keep buying: 60 purchases, most recent the day
+  // before these were written. Everything about its shape is deliberate here,
+  // because the shape is the only part we have evidence for:
+  //
+  //   - one question a buyer actually asks, not an omnibus report
+  //   - three or four steps, all of them cheap
+  //   - €0.20–€0.30, which is the range that has sold; every bundle we price
+  //     at €0.27 or above has sold four times or fewer, and `competitor-
+  //     snapshot` at €1.54 has never sold at all
+  //   - ~1.8× the component sum, so the margin comes from convenience
+  //
+  // The components are the capabilities that actually earn: google-search,
+  // serp-analyze, email-validate, email-deliverability-check, keyword-suggest
+  // and tech-stack-detect were six of the seven biggest external earners over
+  // the 30 days to 2026-08-16, and the first six of those recorded zero
+  // failures in that window. Bundling raises revenue per call about 6.7×
+  // (email-validate earned €11.01 from 368 calls; lead-email-verify earned
+  // €12.00 from 60), which is the entire reason these exist.
+  //
+  // A fifth was designed and dropped before writing: a sitemap/crawl bundle.
+  // `sitemap-parse` takes a sitemap URL (`…/sitemap.xml`), not a site URL, so
+  // it cannot be fed from a plain `url` input without a discovery step we do
+  // not have. Better to ship four that work.
+  // ─────────────────────────────────────────────────────────────────────────
+
+  // ── Competitor Read ──
+  {
+    slug: "competitor-read",
+    name: "Competitor Read",
+    marketingName: "Competitor Read",
+    description:
+      "What is this competitor running, and how do they present themselves? Tech stack, page metadata, and social preview in one call.",
+    longDescription:
+      "Detects the CMS, hosting, analytics and frontend framework behind a competitor's site, extracts their page title, description and structured data, and checks how the page renders when shared on social platforms. The quick competitive read an agent can run on any URL before recommending positioning or a content angle.",
+    agentDescription:
+      "what tech does this competitor use, competitor website analysis, what is this site built with, competitor tech stack and messaging, analyse a competitor page",
+    category: "data-research",
+    priceCents: 20,
+    componentSumCents: 11,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "us-global",
+    targetAudience:
+      "Developers building competitive-research, positioning, or content-strategy agents",
+    transparencyTag: "algorithmic",
+    extendsWith: ["seo-audit", "pricing-page-extract", "brand-mention-search"],
+    inputSchema: {
+      type: "object",
+      properties: { url: { type: "string", format: "uri" } },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://example.com" },
+    exampleOutput: {
+      cms: "WordPress",
+      frontend_framework: "React",
+      title: "Example — the thing we make",
+      description: "One sentence about the product.",
+      og: { image: "https://example.com/og.png" },
+    },
+    // Every step reads $input.url, so all three run together.
+    steps: [
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "meta-extract",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "og-image-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+    ],
+  },
+
+  // ── Page SEO Check ──
+  {
+    slug: "page-seo-check",
+    name: "Page SEO Check",
+    marketingName: "Page SEO Check",
+    description:
+      "Is this page technically fit to rank? Metadata, social preview, load performance, and availability in one call.",
+    longDescription:
+      "Checks whether the page is reachable at all, then the three things that decide whether it can rank before its content is even considered: whether the title, description and structured data are present and well-formed; whether the social preview renders; how fast the page loads; and whether it is reachable at all. Returns a per-area grade so an agent can recommend a specific fix rather than a general one.",
+    agentDescription:
+      "is this page SEO ready, on-page SEO check, why is this page not ranking, check page metadata and speed, technical SEO for a URL",
+    category: "data-research",
+    priceCents: 30,
+    componentSumCents: 16,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "us-global",
+    targetAudience: "Developers building SEO audit, content-ops, or site-monitoring agents",
+    transparencyTag: "algorithmic",
+    extendsWith: ["seo-audit", "backlink-check", "redirect-trace"],
+    inputSchema: {
+      type: "object",
+      properties: { url: { type: "string", format: "uri" } },
+      required: ["url"],
+    },
+    exampleInput: { url: "https://example.com/pricing" },
+    exampleOutput: {
+      title: "Pricing — Example",
+      description: "Plans and pricing.",
+      performance_score: 87,
+      grade: "B",
+      is_up: true,
+      status_code: 200,
+    },
+    steps: [
+      // Availability runs alone and first. There is no conditional-step
+      // support, so an unreachable page still runs the other three and still
+      // bills — but ordering it first guarantees `is_up`/`status_code` are in
+      // the merged result rather than racing three slower checks that will
+      // all fail. The wider gap (skip-and-refund on an unreachable target) is
+      // a platform limitation, recorded in the PR, not something a bundle
+      // definition can fix.
+      {
+        capabilitySlug: "url-health-check",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "meta-extract",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "og-image-check",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+      {
+        capabilitySlug: "page-speed-test",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { url: "$input.url" },
+      },
+    ],
+  },
+
+  // ── Prospect Brief ──
+  {
+    slug: "prospect-brief",
+    name: "Prospect Brief",
+    marketingName: "Prospect Brief",
+    description:
+      "Profile the domain behind a prospect's email: address validity, mail-server setup, domain reputation, and the tech stack running on that domain.",
+    longDescription:
+      "Starts from a single email address: validates the mailbox, then uses its domain to check SPF/DKIM/DMARC configuration, score the domain's reputation, and detect what that domain's site is built with. The step up from Lead Email Verify for agents qualifying a domain, not just an address. Note that it profiles whatever domain the address carries — a consumer mailbox (gmail.com, outlook.com) returns that provider's setup, not an employer's; the is_disposable and is_role_address flags from step 1 are returned so the caller can tell the difference.",
+    agentDescription:
+      "tell me about this prospect from their email, enrich a lead from an email address, qualify a prospect company, what does this company use, research a prospect domain",
+    category: "sales-outreach",
+    priceCents: 30,
+    componentSumCents: 16,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "us-global",
+    targetAudience: "Developers building outbound sales, lead qualification, or CRM enrichment agents",
+    transparencyTag: "algorithmic",
+    extendsWith: ["company-enrich", "domain-reputation", "brand-mention-search"],
+    inputSchema: {
+      type: "object",
+      properties: { email: { type: "string", format: "email" } },
+      required: ["email"],
+    },
+    exampleInput: { email: "sales@example.com" },
+    exampleOutput: {
+      valid: true,
+      domain: "example.com",
+      dmarc: "pass",
+      grade: "A",
+      reputation_score: 88,
+      cms: "Webflow",
+    },
+    // Steps 2-4 all key off the domain that step 1 resolves, exactly as
+    // lead-email-verify does — that chain is the proven part.
+    steps: [
+      {
+        capabilitySlug: "email-validate",
+        stepOrder: 1,
+        canParallel: false,
+        parallelGroup: null,
+        inputMap: { email: "$input.email" },
+      },
+      {
+        capabilitySlug: "email-deliverability-check",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "domain-reputation",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+      {
+        capabilitySlug: "tech-stack-detect",
+        stepOrder: 4,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { domain: "$steps[0].domain" },
+      },
+    ],
+  },
+
+  // ── Keyword Scout ──
+  {
+    slug: "keyword-scout",
+    name: "Keyword Scout",
+    marketingName: "Keyword Scout",
+    description:
+      "Is this keyword worth writing about, and who am I up against? Suggestions, live results, and SERP structure in one call.",
+    longDescription:
+      "Expands a seed keyword into long-tail variants and the questions people actually ask, pulls the live search results, and breaks down the SERP — featured snippets, people-also-ask, and which result types dominate. Everything an agent needs to decide whether a term is worth a piece of content and what shape that content has to take.",
+    agentDescription:
+      "should I target this keyword, keyword research, who ranks for this term, SERP analysis for a keyword, find long tail keywords and competitors",
+    category: "data-research",
+    priceCents: 55,
+    componentSumCents: 28,
+    valueTier: "data-lookup",
+    maintenanceLevel: "near-zero",
+    geography: "us-global",
+    targetAudience: "Developers building SEO research, content-planning, or editorial-calendar agents",
+    transparencyTag: "algorithmic",
+    extendsWith: ["backlink-check", "brand-mention-search", "seo-audit"],
+    // One input on purpose. `country` and `language` were drafted as optional
+    // passthroughs, then removed: the executor resolves an omitted $input
+    // field to `null` and hands it to the capability regardless
+    // (solution-executor.ts — "the downstream capability decides whether a
+    // null input is acceptable"), and none of these three has been tested
+    // with a null locale. Sending an untested null into three paid steps
+    // risks a half-failed call the caller still pays for. Locale targeting
+    // goes back in when somebody asks for it and it has been exercised.
+    inputSchema: {
+      type: "object",
+      properties: { keyword: { type: "string" } },
+      required: ["keyword"],
+    },
+    exampleInput: { keyword: "agent payments" },
+    exampleOutput: {
+      suggestions: ["agent payments api", "ai agent payment rails"],
+      questions: ["how do ai agents pay for things"],
+      total_results_estimate: 41800,
+      featured_snippet: null,
+      serp_features: ["people_also_ask", "organic"],
+    },
+    steps: [
+      {
+        capabilitySlug: "keyword-suggest",
+        stepOrder: 1,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { keyword: "$input.keyword" },
+      },
+      {
+        capabilitySlug: "google-search",
+        stepOrder: 2,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { query: "$input.keyword" },
+      },
+      {
+        capabilitySlug: "serp-analyze",
+        stepOrder: 3,
+        canParallel: true,
+        parallelGroup: 1,
+        inputMap: { keyword: "$input.keyword" },
+      },
+    ],
+  },
+];
